Revert "Leaf 28 interface for database managment with composable for desktop"
