+++ .github/workflows/docker-image.yml
@@ -1,0 +1,44 @@
+name: Docker Image CI/CD
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build-and-push:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+        
+      - name: Log in to Docker Hub
+        uses: docker/login-action@v2
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Build and push server image
+        run: |
+          docker build -t natasadju/leaflink:server-latest -f ./backend/Dockerfile ./backend
+          docker push natasadju/leaflink:server-latest
+
+      - name: Build and push client image
+        run: |
+          docker build -t natasadju/leaflink:frontend-latest -f ./frontend/Dockerfile ./frontend
+          docker push natasadju/leaflink:frontend-latest
+
+  notify:
+    runs-on: ubuntu-latest
+    needs: build-and-push
+    steps:
+      - name: Webhook Action
+        uses: joelwmale/webhook-action@2.4.1
+        with:
+          url: ${{ secrets.WEBHOOK_URL }}
+          headers: '{"Content-Type": "application/json", "X-Hub-Signature-256": "${{ secrets.WEBHOOK_SECRET }}"}'
+          body: '{"text": "Docker images have been successfully built and pushed!"}'
