+++ backend/Dockerfile
@@ -7,3 +7,6 @@ COPY package*.json ./
 RUN npm install
 
 COPY . .
+
+EXPOSE 3000
+CMD [ "node", "app.js" ]
