+++ backend/controllers/eventController.js
@@ -1,4 +1,5 @@
-var EventModel = require('../models/eventModel.js');
+const EventModel = require('../models/eventModel.js');
+const ParkModel = require('../models/parkModel.js');
 
 /**
  * eventController.js
@@ -10,121 +11,120 @@ module.exports = {
     /**
      * eventController.list()
      */
-    list: function (req, res) {
-        EventModel.find(function (err, events) {
-            if (err) {
-                return res.status(500).json({
-                    message: 'Error when getting event.',
-                    error: err
-                });
-            }
-
-            return res.json(events);
-        });
+    list: async (req, res) => {
+        try {
+            const events = await EventModel.find().populate('date'); 
+            res.json(events);
+        } catch (err) {
+            res.status(500).json({
+                message: 'Error when getting events.',
+                error: err.message
+            });
+        }
     },
 
     /**
      * eventController.show()
      */
-    show: function (req, res) {
-        var id = req.params.id;
+    show: async (req, res) => {
+        const id = req.params.id;
 
-        EventModel.findOne({_id: id}, function (err, event) {
-            if (err) {
-                return res.status(500).json({
-                    message: 'Error when getting event.',
-                    error: err
-                });
-            }
-
+        try {
+            const event = await EventModel.findById(id);
             if (!event) {
-                return res.status(404).json({
-                    message: 'No such event'
-                });
+                return res.status(404).json({ message: 'No such event' });
             }
-
-            return res.json(event);
-        });
+            res.json(event);
+        } catch (err) {
+            res.status(500).json({
+                message: 'Error when getting event.',
+                error: err.message
+            });
+        }
     },
 
     /**
      * eventController.create()
      */
-    create: function (req, res) {
-        var event = new EventModel({
-			name : req.body.name,
-			location : req.body.location,
-			date : req.body.date,
-			organizer : req.body.organizer,
-			attendees : req.body.attendees
-        });
-
-        event.save(function (err, event) {
-            if (err) {
-                return res.status(500).json({
-                    message: 'Error when creating event',
-                    error: err
-                });
+    create: async (req, res) => {
+        try {
+            const park = await ParkModel.findById(req.body.location); // Use _id of the park as location
+            if (!park) {
+                return res.status(404).json({ message: 'Park not found' });
             }
 
-            return res.status(201).json(event);
-        });
-    },
+            const event = new EventModel({
+                name: req.body.name,
+                parkId: req.body.location, // Use _id of the park as location
+                date: req.body.date,
+                description: req.body.description,
+            });
 
+            const savedEvent = await event.save();
+            res.status(201).json(savedEvent);
+        } catch (err) {
+            res.status(500).json({
+                message: 'Error when creating event',
+                error: err.message
+            });
+        }
+    },
     /**
      * eventController.update()
      */
-    update: function (req, res) {
-        var id = req.params.id;
+    update: function(req, res){
+        // Add validation rules
+        body('name').optional().notEmpty().withMessage('Name is required'),
+        body('location').optional().notEmpty().withMessage('Location is required'),
+        body('date').optional().notEmpty().withMessage('Date is required'),
+        body('description').optional().notEmpty().withMessage('Description is required'),
 
-        EventModel.findOne({_id: id}, function (err, event) {
-            if (err) {
-                return res.status(500).json({
-                    message: 'Error when getting event',
-                    error: err
-                });
+        async (req, res) => {
+            const id = req.params.id;
+
+            // Handle validation errors
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                return res.status(400).json({ errors: errors.array() });
             }
 
-            if (!event) {
-                return res.status(404).json({
-                    message: 'No such event'
-                });
-            }
-
-            event.name = req.body.name ? req.body.name : event.name;
-			event.location = req.body.location ? req.body.location : event.location;
-			event.date = req.body.date ? req.body.date : event.date;
-			event.organizer = req.body.organizer ? req.body.organizer : event.organizer;
-			event.attendees = req.body.attendees ? req.body.attendees : event.attendees;
-			
-            event.save(function (err, event) {
-                if (err) {
-                    return res.status(500).json({
-                        message: 'Error when updating event.',
-                        error: err
-                    });
+            try {
+                const event = await EventModel.findById(id);
+                if (!event) {
+                    return res.status(404).json({ message: 'No such event' });
                 }
 
-                return res.json(event);
-            });
-        });
+                event.name = req.body.name || event.name;
+                event.location = req.body.location || event.location;
+                event.date = req.body.date || event.date;
+                event.description = req.body.description || event.description;
+                event.organizer = req.body.organizer || event.organizer;
+
+                const updatedEvent = await event.save();
+                res.json(updatedEvent);
+            } catch (err) {
+                res.status(500).json({
+                    message: 'Error when updating event.',
+                    error: err.message
+                });
+            }
+        }
     },
 
     /**
      * eventController.remove()
      */
-    remove: function (req, res) {
-        var id = req.params.id;
+    remove: async (req, res) => {
+        const id = req.params.id;
 
-        EventModel.findByIdAndRemove(id, function (err, event) {
-            if (err) {
-                return res.status(500).json({
-                    message: 'Error when deleting the event.',
-                    error: err
-                });
-            }
-
-            return res.status(204).json();
-        });
-    }
+        try {
+            await EventModel.findByIdAndRemove(id);
+            res.status(204).json();
+        } catch (err) {
+            res.status(500).json({
+                message: 'Error when deleting the event.',
+                error: err.message
+            });
+        }
+    },
 };
