+++ backend/controllers/userController.js
@@ -33,12 +33,10 @@ const login = async (req, res) => {
 };
 
 const dashboard = async (req, res) => {
-    const luckyNumber = Math.floor(Math.random() * 100);
 
-    res.status(200).json({
-        msg: `Hello, ${req.user.name}`,
-        secret: `Here is your authorized data, your lucky number is ${luckyNumber}`,
-    });
+  res.status(200).json({
+    msg: `Hello, ${req.user.name}`
+  });
 };
 
 const getAllUsers = async (req, res) => {
