+++ backend/models/airQualityModel.js
@@ -1,16 +1,16 @@
 var mongoose = require('mongoose');
-var Schema = mongoose.Schema;
+var Schema   = mongoose.Schema;
 
 var airQualitySchema = new Schema({
-    station: String,
-    pm10: Number,
-    pm25: Number,
-    so2: Number,
-    co: Number,
-    ozon: Number,
-    no2: Number,
-    benzen: Number,
-    timestamp: {type: Date, default: Date.now}
+	'location' : String,
+	'timestamp' : Date,
+	'so2' : Number,
+	'co' : Number,
+	'pm10' : Number,
+	'pm25' : Number,
+	'03' : Number,
+	'no2' : Number,
+	'benzene' : Number
 });
 
 module.exports = mongoose.model('airQuality', airQualitySchema);
