+++ backend/package.json
@@ -6,6 +6,10 @@
     "dev": "nodemon ./bin/www"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/lab": "^5.0.0-alpha.170",
+    "@mui/material": "^5.15.18",
     "axios": "^1.6.8",
     "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
@@ -15,13 +19,17 @@
     "dotenv": "^16.4.5",
     "express": "~4.16.1",
     "express-async-errors": "^3.1.1",
+    "express-validator": "^7.1.0",
     "hbs": "~4.0.4",
     "http-errors": "~1.6.3",
+    "install": "^0.13.0",
     "jsonwebtoken": "^9.0.2",
+    "jwt-decode": "^4.0.0",
     "leaflet": "^1.9.4",
     "mongodb": "^6.6.1",
     "mongoose": "^6.9.0",
     "morgan": "~1.9.1",
+    "npm": "^10.8.0",
     "osmtogeojson": "^3.0.0-beta.5",
     "react-leaflet": "^4.2.1",
     "ws": "^8.17.0"
