+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
       dockerfile: Dockerfile
     image: server-side
     container_name: backend-server
-    command: npm start
+    command: node app.js
     volumes:
       - ./backend:/usr/src/app
       - /usr/src/app/node_modules
@@ -16,8 +16,10 @@ services:
     environment:
       - NODE_ENV=development
       - MONGODB_URI=mongodb+srv://leafadmin:leaf123@leafy.gnw7mw8.mongodb.net/leafCollection
+      - JWT_SECRET=9B2F3C4E1A6D8F0B5E7A2C4D9F1E3B8A7D6F4C3E8A2B1C7D4F8E2A6B3C1D9F2
     networks:
       - app-network
+
   frontend:
     build:
       context: ./frontend
@@ -25,8 +27,7 @@ services:
     image: frontend
     container_name: frontend-client
     ports:
-      - "5173:5173"
-
+      - "5173:5173"  # Ensure the port mapping matches the serve command
     networks:
       - app-network
