+++ frontend/package.json
@@ -10,12 +10,20 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@fortawesome/react-fontawesome": "^0.2.1",
+    "@splidejs/splide": "^4.1.4",
     "antd": "^5.17.0",
     "axios": "^1.6.8",
+    "bootstrap": "^5.3.3",
+    "d3": "^7.9.0",
     "leaflet": "^1.9.4",
     "nodemon": "^3.1.0",
     "osmtogeojson": "^3.0.0-beta.5",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.10.2",
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.0",
     "react-leaflet": "^4.2.1",
