+++ frontend/src/pages/Dashboard.jsx
@@ -1,47 +1,44 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react';
 import "../styles/Dashboard.css";
-import { Link, useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import axios from 'axios';
+import Navbar from "./NavBar.jsx";
+import Parks from './Parks.jsx';
+import Events from './Events.jsx';
+
 
 const Dashboard = () => {
-  const [ token, setToken ] = useState(JSON.parse(localStorage.getItem("auth")) || "");
-  const [ data, setData ] = useState({});
-  const navigate = useNavigate();
-
+  const [token, setToken] = useState(JSON.parse(localStorage.getItem("auth")) || "");
+  const [data, setData] = useState({});
+  
   const fetchToken = async () => {
-
     let axiosConfig = {
       headers: {
         'Authorization': `Bearer ${token}`
-    }
+      }
     };
 
     try {
-      const response = await axios.get("http://localhost:5173/api/v1/dashboard", axiosConfig);
+      const response = await axios.get("http://172.211.85.100:5173/api/v1/dashboard", axiosConfig);
       setData({ msg: response.data.msg, luckyNumber: response.data.secret });
     } catch (error) {
       toast.error(error.message);
     }
   }
 
-
-  
   useEffect(() => {
     fetchToken();
-    if(token === ""){
-      navigate("/login");
-      toast.warn("Please login first to access dashboard");
-    }
-  }, [token]);
+  }, []);
 
   return (
-    <div className='dashboard-main'>
-      <h1>Welcome to LeafLink!</h1>
-      <p>You successfully logged in!</p>
-      <Link to="/logout" className="logout-button">Logout</Link>
+    <div>
+      <Navbar />
+      <div className='dashboard-main'>
+        <h1>Welcome to LeafLink!</h1>
+        <Events />
+      </div>
     </div>
-  )
+  );
 }
 
-export default Dashboard
+export default Dashboard;
