+++ frontend/src/pages/Logout.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react'
 import "../styles/Logout.css";
 import { useNavigate } from 'react-router-dom';
+import Navbar from './NavBar.jsx';
 
 const Logout = () => {
 
@@ -14,10 +15,13 @@ const Logout = () => {
     }, []);
 
   return (
-    <div className='logout-main'>
-    <h1>Thank you for saving enviorment!</h1>
-    <p>You logged out! You will be redirected to page where you can login or register</p>
-  </div>
+    <div>
+      <Navbar />
+      <div className='logout-main'>
+        <h1>Thank you for saving enviorment!</h1>
+        <p>You logged out! You will be redirected to page where you can login or register</p>
+      </div>
+    </div>
   )
 }
