+++ frontend/src/styles/Events.css
@@ -1,0 +1,29 @@
+.cards-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+}
+
+.map-box{
+    width: 100%;
+}
+
+.card {
+    background-color: #f5f5f5;
+    border-radius: 8px;
+    padding: 20px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    width: 60%;
+    margin-top: 5%;
+    width: 100%;
+    /* Adjust the width as needed based on your layout */
+}
+
+.card h2 {
+    font-size: 18px;
+    margin-bottom: 10px;
+}
+
+.card p {
+    margin: 0;
+}
