+++ frontend/src/styles/NavBar.css
@@ -1,0 +1,105 @@
+/* Basic styles */
+#navbarSupportedContent ul {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+  }
+  
+  #navbarSupportedContent ul li {
+    display: inline-block;
+  }
+  
+  #navbarSupportedContent ul li a {
+    color: rgba(255, 255, 255, 0.5);
+    text-decoration: none;
+    font-size: 15px;
+    display: flex;
+    align-items: center;
+    padding: 10px 15px; /* Default padding */
+    position: relative;
+    transition: color 0.3s ease-in-out; /* Add transition for color change */
+  }
+  
+  #navbarSupportedContent ul li.active a {
+    color: white; /* Change color for active item */
+  }
+  
+  /* Pulse animation */
+  @keyframes pulse {
+    0% {
+      transform: scale(1);
+    }
+    50% {
+      transform: scale(1.1);
+    }
+    100% {
+      transform: scale(1);
+    }
+  }
+  
+  #navbarSupportedContent ul li.active a {
+    animation: pulse 1s 1; /* Make pulse animation last 1 second and play only once */
+  }
+  
+  /* Add hover effect for nav items */
+  #navbarSupportedContent ul li a:hover {
+    color: white;
+  }
+  
+  body {
+    font-family: 'Roboto', sans-serif;
+  }
+  
+  * {
+    margin: 0;
+    padding: 0;
+  }
+  
+  i {
+    margin-right: 10px;
+  }
+  
+  /*----------bootstrap-navbar-css------------*/
+  .navbar-logo {
+    padding: 15px;
+    color: #fff;
+    font-size: 30px;
+  }
+  
+  .navbar-mainbg {
+    background-color: #7dd181;
+    padding: 0px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  
+  /* Responsive styles */
+  @media only screen and (max-width: 992px) {
+    /* Styles for medium-sized screens (tablets) */
+    #navbarSupportedContent ul li a {
+      padding: 10px; 
+      font-size: 14px; 
+    }
+    
+    .navbar-logo {
+      font-size: 24px; 
+    }
+  }
+  
+  @media only screen and (max-width: 768px) {
+    /* Styles for small screens (mobile devices) */
+    #navbarSupportedContent ul li a {
+      padding: 5px; 
+      font-size: 12px; 
+    }
+    
+    .navbar-logo {
+      font-size: 20px; 
+    }
+  
+    #navbarSupportedContent ul li a span {
+      display: none;
+    }
+  }
+
